🍻 refactor: todo 로직 폴더 구조 변경
